==> scripts/00_automate_cdc_download.ipynb <==
# Automate the Download of CDC WONDER Extractions

This script auto-downloads pre-specified (saved) CDC WONDER queries. The links to the queries must be provided in a .txt file called `cdc_wonder_links.txt` in the appropriate data directory, with line breaks between each query. 

Requires `selenium` and Google Chrome. Tested using MacOS. 

Adapted from `https://github.com/MD-AHTESHAMUDDIN/CDC-Wonder-API`.

from selenium import webdriver
from selenium.webdriver.support.ui import Select

from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By

import time
import os

cwd = os.getcwd()
dir = os.path.dirname(cwd) 

## PATHS TO SET UP / CHANGE
DATA = dir + '/data'
PATH_TO_LINKS = f"{DATA}/cdc_wonder_links_regression_spec.txt"

chromeOptions = webdriver.ChromeOptions()
prefs = {"download.default_directory" : DATA}
chromeOptions.add_experimental_option("prefs",prefs)

## PATH TO 
file_path = PATH_TO_LINKS
with open(file_path, 'r') as file:
    links = file.readlines()
    
https_links = [link.strip() for link in links if link.strip().startswith('https')]

file_path = PATH_TO_LINKS
with open(file_path, 'r') as file:
    links = file.readlines()

https_links = [link.strip() for link in links if link.strip().startswith('https')]

for link in https_links:
    driver = webdriver.Chrome()
    driver.get(link)

    time.sleep(5) # Time delay for the web page to load before accessing the frame

    # Find and click the "I Agree" button
    driver.find_element("xpath", ".//input[@name='action-I Agree']").click()
    time.sleep(5) # Time delay for the web page to load before accessing the frame

    # Find and click the "Send" button -- required if saved at extraction specification
    # THE FOLLOWING TWO LINES SHOULD BE DELETED IF THE SAVED LINK IS TABULAR DATA
    driver.find_element("xpath", ".//input[@value='Send']").click()
    time.sleep(5) # Time delay for the web page to load before accessing the frame

    # Find and click the "Export" button
    driver.find_element("xpath", ".//input[@value='Export']").click()

    time.sleep(25)

    # Close the browser
    driver.quit()
